requirements.txt
altair==5.5.0
attrs==25.3.0
azure-core==1.35.0
azure-iot-device==2.14.0
azure-iot-hub==2.6.1
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
deprecation==2.1.0
fonttools==4.59.0
gitdb==4.0.12
GitPython==3.1.45
idna==3.10
isodate==0.7.2
janus==2.0.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.5
msrest==0.7.1
narwhals==2.0.1
numpy==2.3.2
oauthlib==3.3.1
packaging==25.0
paho-mqtt==1.6.1
pandas==2.3.1
pillow==11.3.0
plotly==6.2.0
protobuf==6.31.1
pyarrow==21.0.0
pydeck==0.9.1
pyparsing==3.2.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
requests-unixsocket2==1.0.0
rpds-py==0.26.0
six==1.17.0
smmap==5.0.2
streamlit==1.47.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
typing_extensions==4.14.1
tzdata==2025.2
uamqp==1.6.11
urllib3==2.5.0
watchdog==6.0.0
